Fix mypy errors in convert, cutout, era5 and orientation
Tighten return types, narrow unions via assert/cast, swap np.where for
xr.where to preserve DataArray typing, and restructure local rebinds to
avoid reassignment conflicts.

+++ atlite/cutout.py
@@ -317,7 +317,7 @@ class Cutout:
     def dx(self) -> float:
         """Spatial resolution on the x coordinates."""
         x = self.coords["x"]
-        return round((x[-1] - x[0]).item() / (x.size - 1), 8)  # type: ignore[no-any-return]
+        return float(round((x[-1] - x[0]).item() / (x.size - 1), 8))
 
     @property
     def dy(self) -> float:
